Use a true stable hash; CPython's isn't.
Don't know how I could have overlooked that fact CPython's is only stable
within the same process.

+++ nvm/libex.py
@@ -32,7 +32,7 @@ pmemobj_structs = """
         PObjPtr ma_values;
         } PDictObject;
     typedef struct {
-        ssize_t me_hash;
+        uint64_t me_hash;
         PObjPtr me_key;
         PObjPtr me_value;           /* Only meaningful for combined table */
         } PDictKeyEntry;
